put methods into SyntaxTree
